requirements.txt
pillow
torchvision
pyyaml
pytorch_lightning
timm
nltk
requests